Move decommissioning_demand function
